+++ src/main/java/pl/grsrpg/player/GamePlayerWarrior.java
@@ -6,10 +6,10 @@ import pl.grsrpg.manager.fight.WarriorFightManager;
 import pl.grsrpg.utils.DiceRoll;
 
 public class GamePlayerWarrior extends GamePlayer {
-    private static final float startMaxHealth = 20;
-    private static final float startStrength = 5;
-    private static final float startAgility = 2;
-    private static final float startMagicPoints = 2;
+    private static final int startMaxHealth = 20;
+    private static final int startStrength = 5;
+    private static final int startAgility = 2;
+    private static final int startMagicPoints = 2;
     private static final int startEquipmentCapacity = 10;
 
     public GamePlayerWarrior(String name, Field currentField) {
@@ -17,19 +17,19 @@ public class GamePlayerWarrior extends GamePlayer {
         this.fightManager = new WarriorFightManager();
     }
 
-    public static float getStartMaxHealth() {
+    public static int getStartMaxHealth() {
         return startMaxHealth;
     }
 
-    public static float getStartStrength() {
+    public static int getStartStrength() {
         return startStrength;
     }
 
-    public static float getStartAgility() {
+    public static int getStartAgility() {
         return startAgility;
     }
 
-    public static float getStartMagicPoints() {
+    public static int getStartMagicPoints() {
         return startMagicPoints;
     }
